choose_disk: Adding "ceph_type" support
ceph_type is about defining the kind of disk defined by profile.
It could be "data" or "journal".

The output is also changed to be more Ansible-like compliant and will be used
directly to call ceph-disk.

Note the new output doesn't report the non-configured device to avoid adding more logic
into the scenario.