ceph-mon: fix support for ipv6 on containerized mons
e8187f6 does not fix the ipv6 as expected since `ansible_default_*` are
filled with the IP address carried by the network interface used by the
default gateway route. By the way it assumes that the MON_IP address will
be this IP address which is not always the case.

We need to keep using the previous fact but add some intelligency in the
template to determine how to retrieve the ipv4|ipv6 address since the path
to the fact in `hostvars` is not the same according to ipv4 vs ipv6 case.

Signed-off-by: Guillaume Abrioux <gabrioux@redhat.com>